requirements.txt
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
common-toolbox==0.0.16
DBUtils==1.2
gevent==1.2.2
greenlet==0.4.13
idna==2.6
mock==2.0.0
nose==1.3.7
numpy==1.13.3
pandas==0.22.0
pbr==3.1.1
pika==0.11.2
Pillow==5.0.0
pymongo==3.5.1
PyMySQL==0.8.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
scipy==1.0.0
six==1.11.0
SQLAlchemy==1.2.0
tornado==4.5.3
urllib3==1.22
vine==1.1.4
